requirements.txt
streamlit
langchain
langchain-community
langchain-groq
chromadb
pandas
python-dotenv
unstructured
selenium